pyproject.toml
[tool.poetry]
name = "ecom"
version = "0.1.0"
description = "A e-commerce platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2.7"
asgiref = "3.7.2"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
colorama = "0.4.6"
cryptography = "41.0.5"
defusedxml = "0.7.1"
dj-rest-auth = "5.0.1"
django-allauth = "0.58.1"
django-js-asset = "2.1.0"
django-mptt = "0.15.0"
django-rest-auth = "0.9.5"
djangorestframework = "3.14.0"
idna = "3.4"
iniconfig = "2.0.0"
oauthlib = "3.2.2"
packaging = "23.2"
Pillow = "10.1.0"
pluggy = "1.3.0"
pycparser = "2.21"
PyJWT = "2.8.0"
pytest = "7.4.2"
pytest-django = "4.6.0"
python-dotenv = "1.0.0"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
six = "1.16.0"
sqlparse = "0.4.4"
tzdata = "2023.3"
urllib3 = "2.0.7"
djangorestframework-simplejwt = "^5.3.0"
flake8 = "6.1.0"
pre-commit = "3.4.0"
pylint = "3.0.1"
djlint = "1.34.0"
commitizen = "3.12.0"
black = "^23.10.1"
autopep8 = "^2.0.4"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
bandit = "^1.7.5"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[tool.black]
line-length = 120
include = '\.py?$'
exclude = 'settings.py'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[tool.djlint]
profile = "django"
ignore = "H031"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
